+++ R/wkt.R
@@ -82,9 +82,14 @@ prnt.GEOMETRYCOLLECTION = function(x, ..., EWKT = TRUE) {
 #' @param ... modifiers; in particular \code{digits} can be passed to control the number of digits used
 #' @name st_as_text
 #' @details The returned WKT representation of simple feature geometry conforms to the
-#' \href{https://www.ogc.org/standard/sfa/}{simple features access} specification and extensions
+#' \href{https://www.ogc.org/publications/standard/sfa/}{simple features access} specification and extensions
 #' (known as EWKT, supported by PostGIS and other simple features implementations for addition of 
 #' a SRID to a WKT string).
+#' @note To improve conversion performance, the lwgeom package can be used (it must be installed
+#' beforehand) and set the \code{Sys.setenv("LWGEOM_WKT" = "true")} environment variable. This
+#' will also result in faster printing of complex geometries. Note that the representation as WKT is
+#' different from the sf package and may cause reproducibility problems. An alternative solution is
+#' to use the [lwgeom::st_astext()] or [wk::as_wkt()] functions.
 #'
 #' @export
 st_as_text = function(x, ...) UseMethod("st_as_text")
@@ -139,7 +144,7 @@ st_as_text.sfc = function(x, ..., EWKT = FALSE) {
 #' @rdname st_as_sfc
 #' @md
 #' @details If `x` is a character vector, it should be a vector containing
-#' [well-known-text](https://www.ogc.org/standard/wkt-crs/), or
+#' [well-known-text](https://www.ogc.org/publications/standard/wkt-crs/), or
 #' Postgis EWKT or GeoJSON representations of a single geometry for each vector element.
 #' @param crs integer or character; coordinate reference system for the
 #' @param GeoJSON logical; if \code{TRUE}, try to read geometries from GeoJSON text strings